+++ SoftwareRenderer/object_handler.h
@@ -7,34 +7,34 @@
 class Scalable
 {
 public:
-	Vector<double, 3> Scale = Vector<double, 3>({1.0, 1.0, 1.0});
+	Vector<float, 3> Scale = Vector<float, 3>({1.0, 1.0, 1.0});
 };
 
 class ObjectData
 {
 public:
-	Vector<double, 3> Origin = Vector<double, 3>({0.0, 0.0, 0.0});
-	Vector<double, 3> Rotation = Vector<double, 3>({0.0, 0.0, 0.0});
+	Vector<float, 3> Origin = Vector<float, 3>({0.0, 0.0, 0.0});
+	Vector<float, 3> Rotation = Vector<float, 3>({0.0, 0.0, 0.0});
 };
 
 class Camera : public ObjectData
 {
 public:
-	double FocalLength;
+	float FocalLength;
 
-	void ShiftFocalLength(double change)
+	void ShiftFocalLength(float change)
 	{
 		FocalLength += change;
 	}
 
-	void ShiftLocation(Vector<double, 3> shift)
+	void ShiftLocation(Vector<float, 3> shift)
 	{
 		Origin[0] += shift[0];
 		Origin[1] += shift[1];
 		Origin[2] += shift[2];
 	}
 
-	void ShiftRotation(Vector<double, 3> rotate)
+	void ShiftRotation(Vector<float, 3> rotate)
 	{
 		Rotation[0] += rotate[0];
 		Rotation[1] += rotate[1];
@@ -45,12 +45,12 @@ public:
 class Model : public ObjectData, public Scalable
 {
 	int LoadObject(std::string path);
-	std::array<double, 16> ModelMatrix;
+	std::array<float, 16> ModelMatrix;
 public:
-	std::vector<Vector<double, 3>> Vertex;
+	std::vector<Vector<float, 3>> Vertex;
 	std::vector<Vector<int, 3>> Face;
 
-	Model(std::string path, Vector<double, 3> Loc = Vector<double, 3>(), Vector<double, 3> Rot = Vector<double, 3>(), Vector<double, 3> Sca = Vector<double, 3>({1,1,1}))
+	Model(std::string path, Vector<float, 3> Loc = Vector<float, 3>(), Vector<float, 3> Rot = Vector<float, 3>(), Vector<float, 3> Sca = Vector<float, 3>({1,1,1}))
 	{
 		Origin = Loc;
 		Rotation = Rot;
@@ -58,6 +58,6 @@ public:
 		this->LoadObject(path);
 	}
 
-	std::array<double, 16> GetModelMatrix();
+	std::array<float, 16> GetModelMatrix();
 };
